Create data_task_1
contains a list of financial instruments. Each has a Nominal, a currency and a maturity. The Nominal shows the amount of a future cash flow and a sign, indicating the direction of the cash flow. If it is negative, our bank pays the counter-party, if it is positive, we will receive the money. The maturity indicates when the financial contract ends and the Nominal gets paid.