Add files via upload
Includes simulated interest rates for a time period of 10 years, e.g. the given value shows the interest rate at that date for all tenors. The rates at a given moment in time are the same for all maturities (curve structure, tenor spreads or forward rates are not taken into account). The values are noted in percent %.